Code cleaning and compatibility with Python3

+++ pbdlib/functions.py
@@ -5,8 +5,8 @@ from scipy.special import gamma, gammaln
 colvec = lambda x: np.array(x).reshape(-1, 1)
 rowvec = lambda x: np.array(x).reshape(1, -1)
 
-realmin = np.finfo(np.float32).tiny
-realmax = np.finfo(np.float32).max
+realmin = np.finfo(np.float64).tiny
+realmax = np.finfo(np.float64).max
 
 def limit_gains(gains, gain_limit):
 	"""
